requirements.txt
# ============================================
# CORE DATA SCIENCE
# ============================================
pandas==2.3.3
numpy==2.1.3
scipy==1.16.3

# ============================================
# MACHINE LEARNING
# ============================================
scikit-learn==1.8.0
xgboost-cpu==3.1.2
lightgbm==4.6.0
joblib==1.5.3

# ============================================
# DEEP LEARNING
# ============================================
tensorflow==2.20.0.dev0+selfbuilt
keras-nightly==3.14.0.dev2025122304

# ============================================
# TIME SERIES
# ============================================
statsmodels==0.14.6
prophet==1.2.1
pmdarima==2.1.1

# ============================================
# VISUALIZATION
# ============================================
matplotlib==3.10.8
seaborn==0.13.2
plotly==6.5.0

# ============================================
# DASHBOARD
# ============================================
streamlit==1.52.2

# ============================================
# INTEL OPTIMIZATIONS (optionnel)
# ============================================
scikit-learn-intelex==2025.10.0

# ============================================
# UTILITIES
# ============================================
tqdm==4.67.1
